+++ src/main/java/com/example/restdocs/controller/HelloController.java
@@ -1,11 +1,12 @@
-package com.example.restdocs;
+package com.example.restdocs.controller;
 
+import com.example.restdocs.controller.response.ApiResult;
+import com.example.restdocs.controller.request.OptionalBody;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
