+++ src/main/java/com/example/restdocs/controller/request/OptionalBody.java
@@ -1,4 +1,4 @@
-package com.example.restdocs;
+package com.example.restdocs.controller.request;
 
 public record OptionalBody(String required, String optional) {
